AmbienteSena/urls.py
# AmbienteSena/urls.py - URLs principales del proyecto

from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path('admin/', admin.site.urls),
    
    # URLs para autenticación JWT
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    
    # ✅ URLs de la API REST bajo /api/
    path('api/', include('cursos.urls')),  # Esto creará /api/cursos/, /api/horarios/, etc.
    
    # ✅ Página principal (opcional)
    # path('', TemplateView.as_view(template_name='index.html'), name='home'),
]

# Servir archivos media en desarrollo
if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)