AmbienteSena/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==24.2
Pillow==10.4.0
setuptools>=69.0.0